+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [unreleased]
+
+### 🚀 Features
+
+- *(cliff)* Add cliff.toml and update solution
 ## [0.2.0.0] - 2025-09-01
 
 ### 🚀 Features